% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_single.R
\name{distance_single}
\alias{distance_single}
\title{Baseline scenario - pairvise origin / destination call}
\usage{
distance_single(origin, destination, mode = "driving", key = get.api.key())
}
\arguments{
\item{origin}{a single origin}

\item{destination}{a single destination}

\item{mode}{mode of tranport}

\item{key}{Google API key}
}
\value{
a list of values: status, duration text + value, distance text + value
}
\description{
Baseline scenario - pairvise origin / destination call
}
