% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmapsdistance-package.R
\name{get.api.key}
\alias{get.api.key}
\title{Get the Google Maps API key}
\usage{
get.api.key()
}
\value{
the user's api key
}
\description{
This function returns the user's Google Maps API key that was defined with
\code{set.api.key}.
}
