% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmapsdistance.R
\name{get_api_key}
\alias{get_api_key}
\title{Get the Google Maps API key}
\usage{
get_api_key()
}
\value{
the user's api key
}
\description{
This function returns the user's Google Maps API key that was defined with
\code{set_api_key}.
}
\examples{
get_api_key()
}
