% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{gmapsdistance2}
\alias{gmapsdistance2}
\alias{gmapsdistance2-package}
\title{Compute Distance with Google Maps}
\description{
Get distance and travel time between two points from Google Maps.
Four possible modes of transportation (bicycling, walking, driving and
public transportation).
}
