% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmapsdistance.R
\docType{data}
\name{pkg.env}
\alias{pkg.env}
\title{Define package environment}
\format{\preformatted{<environment: 0x3f1e250> 
}}
\usage{
pkg.env
}
\description{
\code{pkg.env} is a package environment that contains the variable
\code{api.key} with the user's Google Maps API key
}
\keyword{datasets}

