% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmapsdistance.R
\docType{data}
\name{pkg_env}
\alias{pkg_env}
\title{Define package environment}
\format{An object of class \code{environment} of length 1.}
\usage{
pkg_env
}
\description{
\code{pkg.env} is a package environment that contains the variable
\code{api_key} with the user's Google Maps API key
}
\keyword{datasets}
