% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmapsdistance-package.R
\name{set.api.key}
\alias{set.api.key}
\title{Set the Google Maps API key}
\usage{
set.api.key(key)
}
\arguments{
\item{key}{is the user's Google Maps API key}
}
\description{
This function stores a user's Google Maps API key as the package's
environmental variable
}
\examples{
#DONTRUN
set.api.key("MY-GOOGLE-MAPS-API-KEY")
}
