% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmapsdistance.R
\name{set_api_key}
\alias{set_api_key}
\title{Set the Google Maps API key}
\usage{
set_api_key(key)
}
\arguments{
\item{key}{is the user's Google Maps API key}
}
\description{
This function stores a user's Google Maps API key as the package's
environmental variable
}
\examples{
\dontrun{
set_api_key("MY-GOOGLE-MAPS-API-KEY")
}
}
