% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmapsdistance-package.R
\docType{data}
\name{pkg.env}
\alias{pkg.env}
\title{Define package environment}
\format{
An object of class \code{environment} of length 1.
}
\usage{
pkg.env
}
\description{
\code{pkg.env} is a package environment that contains the variable
\code{api.key} with the user's Google Maps API key
}
\keyword{internal}
